requirements.txt

Flask


psycopg2-binary
pgvector


sentence-transformers
numpy
pandas
scikit-learn
python-dotenv

Background 
celery
redis


llama-index
llama-index-llms-ollama